taelibmodeltrans/requirements.txt
asgiref==3.7.2
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-appconf==1.0.6
django-cors-headers==4.3.1
django-googledrive-storage==1.5.0
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
django-rest-auth==0.9.5
django-rosetta==0.10.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
google-api-core==2.10.2
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
install==1.3.5
mysqlclient==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
pillow==10.2.0
polib==1.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
sqlparse @ file:///work/perseverance-python-buildout/croot/sqlparse_1698874574933/work
tabulate==0.9.0
tqdm==4.66.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.41.2
